requirements.txt
PyOpenGL
glfw
numpy
imgui[glfw]
Pillow
